src/requirements.txt
os
sys
markdown
tkinter
tkcode
pyperclip
Pillow
pyautogui
win32clipboard